Rename package and move `MainMenu` to `ui`
