+++ src/com/osudroid/ui/v2/game/CirclePiece.kt
@@ -1,4 +1,4 @@
-package com.osudroid.ui.v2.hitobjects
+package com.osudroid.ui.v2.game
 
 import com.reco1l.andengine.*
 import com.reco1l.andengine.container.*
