+++ src/ru/nsu/ccfit/zuev/osu/game/GameplaySlider.java
@@ -11,10 +11,10 @@ import com.reco1l.andengine.sprite.AnimatedSprite;
 import com.reco1l.andengine.sprite.ExtendedSprite;
 import com.reco1l.andengine.Modifiers;
 import com.reco1l.andengine.Anchor;
-import com.osudroid.ui.v2.hitobjects.SliderTickSprite;
-import com.osudroid.ui.v2.hitobjects.CirclePiece;
-import com.osudroid.ui.v2.hitobjects.NumberedCirclePiece;
-import com.osudroid.ui.v2.hitobjects.SliderTickContainer;
+import com.osudroid.ui.v2.game.SliderTickSprite;
+import com.osudroid.ui.v2.game.CirclePiece;
+import com.osudroid.ui.v2.game.NumberedCirclePiece;
+import com.osudroid.ui.v2.game.SliderTickContainer;
 import com.rian.osu.beatmap.hitobject.BankHitSampleInfo;
 import com.rian.osu.beatmap.hitobject.Slider;
 import com.rian.osu.beatmap.hitobject.sliderobject.SliderTick;
